+++ documentation/playground/src/base/IconExamples.tsx
@@ -5,7 +5,6 @@ export function IconExamples() {
     <div className="u-margin">
       <h2>Icons</h2>
       <div className="u-margin">
-        <Icon name="alarm-bell-1" />
         <Icon name="alert-diamond" />
         <Icon name="mouse" screenReaderText="Click to connect your mouse" />
         <Icon thin name="ai-mouse" screenReaderText="Click to connect your mouse" />